pyproject.toml
[tool.poetry]
name = "fmp-data"
version = "0.1.0"
description = "Financial Modeling Prep (FMP) API client with data persistence"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "fmp_data", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11, <4.0"
pandas = "^2.1.4"
redis = "^5.0.1"
pyyaml = "^6.0.1"
loguru = "^0.7.2"
rich = "^13.7.0"
motor = "^3.3.2"
pydantic = "^2.5.3"
pytz = "^2023.3"
aiohttp = "^3.11.10"
exchange-calendars = "^4.6"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
pytest-cov = "^4.1.0"
black = "^23.12.1"
isort = "^5.13.2"
mypy = "^1.8.0"
ruff = "^0.1.9"
types-redis = "^4.6.0"
types-PyYAML = "^6.0.12"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = [
    "pandas.*",
    "redis.*",
    "rich.*",
    "loguru.*",
    "motor.*"
]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
log_cli = true
log_cli_level = "DEBUG"
log_cli_format = "%(asctime)s [%(levelname)8s] %(message)s (%(filename)s:%(lineno)s)"
log_cli_date_format = "%Y-%m-%d %H:%M:%S"
